tortuga-api/src/test/java/de/computerstudienwerkstatt/tortuga/controller/DoorAuthorisationAttemptControllerTest.java
package de.computerstudienwerkstatt.tortuga.controller;

import de.computerstudienwerkstatt.tortuga.TestContext;
import de.computerstudienwerkstatt.tortuga.controller.api.statistics.DoorAuthorisationAttemptController;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.SpringApplicationConfiguration;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import org.springframework.test.context.web.WebAppConfiguration;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.setup.MockMvcBuilders;
import org.springframework.web.context.WebApplicationContext;

import javax.annotation.Resource;
import java.lang.reflect.Method;

import static junit.framework.TestCase.assertTrue;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

/**
 * @author Mischa Holz
 */
@RunWith(SpringJUnit4ClassRunner.class)
@SpringApplicationConfiguration(classes = TestContext.class)
@WebAppConfiguration
public class DoorAuthorisationAttemptControllerTest {

    @Resource
    private WebApplicationContext webApplicationContext;

    private MockMvc mockMvc;

    @Autowired
    private DoorAuthorisationAttemptController doorAuthorisationAttemptController;

    @Before
    public void setUp() throws Exception {
        mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
    }

    @Test
    public void testFindAll() throws Exception {
        mockMvc.perform(get("/api/v1/stats/doorauthorisationattempts").contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk());
    }

    @Test
    public void testApiBase() throws Exception {
        Method method = DoorAuthorisationAttemptController.class.getDeclaredMethod("getApiBase");
        method.setAccessible(true);
        Object ret = method.invoke(doorAuthorisationAttemptController);

        assertTrue(ret instanceof String);
        assertTrue(ret.toString().equals(DoorAuthorisationAttemptController.API_BASE));
    }
}
